almost done 75 %

+++ ExamManager/app/__init__.py
@@ -1,14 +1,16 @@
+import os
+rewrite = False
+if os.path.exists('./Exammanager/app.db') and rewrite !=True:
+    os.remove('./Exammanager/app.db')
+    rewrite = True
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import *
-import os
 
 app = Flask(__name__)
 app.config.from_object('config')
 db = SQLAlchemy(app=app)
 
 from app import views, models
-#import app.views
-if os.path.exists('./Exammanager/app.db'):
-    os.remove('./Exammanager/app.db')
-from app import db_create, default, actions
+from app import db_create, default
+from app import actions
